Fix integer truncation in check_cmac_parms() causing a NULL pointer dereference
Function check_cmac_parms() allows message to be NULL, if message_length is
zero. The callers of check_cmac_parms() supply a unsigned long message length
to check_cmac_parms() which truncates the length to an unsigned int.

If an unsigned long message length truncates to 0 as unsigned int (for
example 0x100000000UL) then a NULL message is accepted, although the
subsequent code operate on the non-zero unsigned long message length and
thus attempt to dereference the NULL message.

While at it fix a similar integer truncation with check_aes_parms(). There
it does not have any bad consequences, but it is still worth to fix.

Reported by: Aisle Research

+++ src/ica_api.c
@@ -166,7 +166,7 @@ static unsigned int check_des_parms(unsigned int mode,
 }
 
 static unsigned int check_aes_parms(unsigned int mode,
-				    unsigned int data_length,
+				    unsigned long data_length,
 				    const unsigned char *in_data,
 				    const unsigned char *iv,
 				    unsigned int key_length,
@@ -234,7 +234,7 @@ static unsigned int check_aes_parms(unsigned int mode,
 }
 
 static unsigned int check_cmac_parms(unsigned int block_size,
-				     const unsigned char *message, unsigned int message_length,
+				     const unsigned char *message, unsigned long message_length,
 				     unsigned char *mac, unsigned int mac_length,
 				     unsigned char *keys, unsigned int key_length,
 				     unsigned char *iv)
